@@ -12,9 +12,9 @@
 
     public string LastName { get; set; }
 
-    public IEnumerable<ArtistSlug> Slugs { get; set; }
+    public string FullName { get; set; }
 
-    public string ImageUrl { get; set; }
+    public IEnumerable<ArtistSlug> Slugs { get; set; } = new List<ArtistSlug>();
 
     public bool IsApproved { get; set; }
 
@@ -22,9 +22,11 @@
 
     public IEnumerable<Lyric> Lyrics { get; set; }
 
+    public ArtistImage Image { get; set; }
+
     public DateTime CreatedAt { get; set; }
 
-    public DateTime ModifiedAt { get; set; }
+    public DateTime? ModifiedAt { get; set; }
 
     public bool IsDeleted { get; set; }
   }
