@@ -13,7 +13,7 @@
 
     public DateTime CreatedAt { get; set; }
 
-    public DateTime ModifiedAt { get; set; }
+    public DateTime? ModifiedAt { get; set; }
 
     public bool IsDeleted { get; set; }
 
