@@ -8,7 +8,7 @@
 
     DateTime CreatedAt { get; set; }
 
-    DateTime ModifiedAt { get; set; }
+    DateTime? ModifiedAt { get; set; }
 
     bool IsDeleted { get; set; }
   }
