@@ -1,11 +1,20 @@
 ﻿namespace Bejebeje.Services.Services.Interfaces
 {
-  using System.Collections.Generic;
   using System.Threading.Tasks;
-  using Bejebeje.ViewModels.Artist;
+  using Bejebeje.Models.Artist;
 
   public interface IArtistsService
   {
-    Task<IList<ArtistCardViewModel>> GetArtistsAsync();
+    Task<int> GetArtistIdAsync(string artistSlug);
+
+    Task<bool> ArtistExistsAsync(string artistSlug);
+
+    Task<ArtistDetailsResponse> GetArtistDetailsAsync(string artistSlug);
+
+    Task<PagedArtistsResponse> GetArtistsAsync(int offset, int limit);
+
+    Task<CreateNewArtistResponse> CreateNewArtistAsync(CreateNewArtistRequest request);
+
+    Task<PagedArtistsResponse> SearchArtistsAsync(string artistName, int offset, int limit);
   }
 }
